+++ lib/gradient_container.dart
@@ -1,21 +1,20 @@
 import 'package:flutter/material.dart';
-import 'package:quiz_app/custom_button.dart';
-import 'package:quiz_app/styled_text.dart';
 
 const startAlignment = Alignment.topCenter;
 const endAlignment = Alignment.bottomCenter;
 
 class GradientContainer extends StatelessWidget {
-  const GradientContainer({required colors, super.key})
-      : gradientColors = colors;
+  const GradientContainer(
+      {required this.content, required this.gradientColors, super.key});
 
-  const GradientContainer.home({super.key})
+  const GradientContainer.home({required this.content, super.key})
       : gradientColors = const [
-          Color.fromARGB(255, 15, 0, 153),
+          Color.fromARGB(255, 54, 0, 153),
           Color.fromARGB(255, 50, 0, 92)
         ];
 
   final List<Color> gradientColors;
+  final Widget content;
 
   @override
   Widget build(context) {
@@ -24,26 +23,7 @@ class GradientContainer extends StatelessWidget {
         gradient: RadialGradient(
             colors: gradientColors, center: Alignment.center, radius: 0.9),
       ),
-      child: Center(
-          child: Column(
-        mainAxisSize: MainAxisSize.min,
-        mainAxisAlignment: MainAxisAlignment.center,
-        children: [
-          Image.asset(
-            "assets/images/quiz-logo.png",
-            width: 250,
-          ),
-          const SizedBox(height: 50),
-          const StyledText.title(text: "Learn Flutter the fun way!"),
-          const SizedBox(height: 50),
-          CustomButton(onPressed: startQuiz, 
-          text: const StyledText.button(text: "Start Quiz"))
-        ],
-      )),
+      child: content,
     );
-  }
-
-  void startQuiz() {
-    print("Starting quiz...");
   }
 }
