--- lib/views/start_screen.dart
@@ -1,10 +1,0 @@
-import 'package:flutter/material.dart';
-
-class StartScreen extends StatelessWidget {
-
-  @override
-  Widget build(context) {
-    return 
-  }
-
-}
